storage: explicitly order commands in the command sequencer
The previous command sequencer dataflow relied on the assumption that
updates passing through Timely channels are not reordered. That isn't an
assumption we should make, so this commit adds code to explicitly track
and restore the correct ordering by assigning indexes to transmitted
commands and then making sure that no indexes are skipped downstream.