adapter: Fix startup parameter default values (#31166)
There are some server configuration parameters that we need during
startup, before the catalog has been fully opened. To get the current
value we manually look at the current value in LaunchDarkly (which is
confusingly returned as `None` if it matches the default value in the
code), then we manually look in the catalog, then we manually look in
the server parameter default override CLI arg map.

Previously, for some of the parameters, if we weren't able to find the
value in any of those places, then we hard-code in the value false.
This is not correct, and instead we should fall back to their default
value, which is sometimes true. This commit fixes the issue by falling
back to their default values.