storage: make Kafka sources work with zero exports
Before, when we had zero outputs, we would repeatedly try and create a
partition queue for the same partition, because our `self.last_offsets`
is empty. The existing early exit would never trigger because it only
works when there is _some_ output state.

Now, we always early exit when we know we don't have any outputs.