adapter: Enrich expr cache version in dev (#31233)
The expression cache uses the build version as part of the key to ensure
that a materialize instance only tries to deserialize expressions that
were serialized using the same version of the code. However, a single
dev build version can be used for many different builds, which can lead
to errors in deserialization when running locally. This commit fixes the
issue by enriching dev build versions with build metadata containing the
build ID.

Note: we could enrich all build versions with build ID metadata without
any correctness issues. However, the ID is fairly long and must be
stored for each expression. Since we already have the guarantee that a
prod build version is only ever used with a single build, it would be
wasted space for prod build versions.

Fixes #MaterializeInc/database-issues/issues/8930

---------

Co-authored-by: Dennis Felsing <dennis@felsing.org>