storage: make loadgens export a progress stream
This commit makes the simple load generators export a progress stream,
giving the remap operator the ability to mint bindings earlier, thus
allowing the reclock operator to pass through snapshot updates instead
of stashing them all. The effect of this that load generators now need
constant memory during hydration, instead of memory proportional to the
snapshot size.

This commit also makes the progress-stream non-optional in source
rendering, to avoid pitfalls like this in the future.