0dt: in preflight checks, notice DDL changes and restart read-only envd
Before, a hydrating read-only environment would read a catalog snapshot
on bootstrap and then not subscribe to further changes. This means that
we would not be hydrating collections/replicas that are created in the
old version that is still running, and those would then have to be
hydrated after cutting over to the new version.

With this change, we periodically check if new collections/replicas
where created and we also check right before announcing as ready to
promote. When we _do_ notice there was relevant DDL, we halt. This will
make it so we're restarted in read-only mode again, and can now read an
up-to-date catalog snapshot.

It's important to note that any running clusters are not restarted, so
any work that has already gone into hydration will not be lost.