migration test: Skip invisible columns in old version
The newly introduced migration tests run from the previous Mz version
and test that we can migrate the source when running a new version.
(Does this even make sense or can we just use a new version for both?
Unfortunately Joe left so I can't ask him what the original motivation
for that was)

For now we can skip the td file when it's starting in an older version.

Failure seen in https://buildkite.com/materialize/nightly/builds/11076#0194ddd4-9f47-4888-bbaa-89ddfbc86da8