Switch to upstream launchdarkly sdk
Switch to the upstream launchdarkly-server-sdk crate instead of our own
fork. This comes at the expense of losing some metrics that we added to
our crate but never attempted to upstream. It is not easy to replicate
the metrics without creating our own wrapper of the http library (or
continue to maintain our own fork). I'd rather lose the metrics over
continuing to maintain our own fork because the maintenance will cost time.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>

# Conflicts:
#	Cargo.lock
#	src/balancerd/Cargo.toml