compute/correction-v2: introduce a Stage to speed up small inserts
This commit extends the `CorrectionV2` data structure with a `Stage`
that accumulates inserted updates before they get inserted into the
sorted chains. This significantly reduces the amount of chain merges
that need to be done in workloads that trickle in large amounts of
updates in small batches, greatly speeding up these workloads.