cluster: ensure workers get activated on controller reconnect
This commit fixes a race condition that could cause workers to not get
activated when their connection to the controller dropped. The
`LocalClient` connected to each worker makes sure that an activation is
sent when it drops, but there is a possibility that the controller
connection fails before a `LocalClient` has been created, in which case
any worker activators created so far would be dropped without activating
the worker.

The fix implemented here is to immediately wrap worker activators in a
`LocalActivator` type with a `Drop` impl that ensures activation on
drop. Some care needs to be taken, by ordering declarations
appropriately, to ensure that activators are always dropped _after_ the
corresponding command channels have been dropped, to ensure workers
observe the command channel disconnection when they are activated.