txn-wal fix 1
Previously this PR was panicking on startup inside of TxnsCacheState when validating that
our data write times were fully compacted. Specifically we had two data write times for
the same timestamp, and said timestamp was < unapplied_writes_ts. This makes sense because
a fully compacted list of writes is designed to track only the 'latest'. The bug was that
we committed two batches at the same time, and thus put the same timestamp in the writes
list twice.

We committed two batches at the same time because now we were commits both in-memory Rows
and pre-staged batches to the same table during group commit. This should be a workflow
that is supported, so the fix is when recording the timestamp for a write, only record it
if it's not the most recent entry in our list. This should cover the case of multiple
different batches getting appended with the same timestamp.