txn-wal fix 2
We were seeing a panic in Persist when validating that a batch could be truncated see
(see persist-client::batch::validate_truncate_batch) and similar to txn-wal fix 1, the root
cause was appending two batches at the same time, one from in-memory Rows and another with
pre-staged Batches.

When appending txn-wal would rewrite the timestamp for the pre-staged batches, but not for
the Batches that were newly created from our in-memory Rows. When running the aforementioned
validation check, if _any_ Batch in the current set being appended has a rewritten timestamp
then it would validate there isn't any data to truncate, below the batches lower. Because
the pre-staged Batches had their timestamp re-written, this check was getting run for the
newly formed batches, and failing.

There are two possible fixes:
1. Relax this validation check in persist
2. Update txn-wal to rewrite the timestamp for the batches it creates from in-memory rows

We opted for solution [2] since it results in all batches from txn-wal getting run through
the same code path, and thus not only fixes the bug but IMO makes things easier to reason
about.