[adapter] don't block on builtin table write in `Session` creation (#31309)
This PR does two things:

1. Changes explicitly triggered group commits (e.g. a write to
`mz_sessions`) to have a minimum batching interval. Today when we
trigger a group commit we only allow 1 to be in-flight at a time, this
allows pending writes to naturally batch. But if the write is small
there's a good chance Persist will inline it and it will take low 10s of
milliseconds. Added in this PR is a dyncfg `group_commit_batch_duration`
which ensures we will wait X duration before running the next group
commit. The default value is 100ms.
2. When starting a `Session`/connection we have to update the
`mz_sessions` table. Waiting for this write to complete has been moved
to when a Session receives a query, and if that query reads from the
`mz_sessions` table. This means starting a `Session` is a completely
in-memory operation, and waiting for the builtin table write to complete
happens concurrently with the network roundtrip of responding to the
client, and them sending their request.

Theoretically it could _increase_ latency for opening a new connection,
if that connection runs `SELECT ... FROM mz_sessions` because we now we
might wait up to 100ms for the write to complete. But this seems like an
acceptable tradeoff because it speeds up all other queries and is
tune-able in LaunchDarkly if it becomes an issue.

### Motivation

Fixes https://github.com/MaterializeInc/database-issues/issues/8951

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [x] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [x] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [x] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.