storage: eagerly drop source/collection state in drop_sources_unvalidated
Before, we were waiting on DroppedId messages (for ingestions) or
PendingCompactionCommand (for tables and webhook sources) for a couple
of things:
- cleaning up collection state in the controller
- updating mz_storage_shards (the global id -> shard id mapping)
- updating source/sink status to "dropped"
- removing statistics tracking

In some situations we don't get DroppedId messages, so we would not do
these updates/cleanups. The common of these situations are:
- dropping a replica right after dropping sources/sinks
- dropping a whole cluster

Now, we eagerly apply updates to our internal collections and also
eagerly clean up state.

The benefits of this change:
- less code
- control flow of dropping things is more localized and easier to
  understand, I challenge folks to understand how dropping ingestions
  worked before this change ... ;-)
- we fix problems where we inadvertently didn't drop state and update
  internal collections

One thing we lose with this change: we don't assert that
FrontierResponses (from the cluster/replica) match an entry in our
state. This is because we eagerly clean out that state. What we _could_
do is introduce an in-memory dropped_ids field, to which we add IDs when
dropping. And then when we receive a spurious message from the cluster
we can check it against that.