adapter: downgrade read-holds in read-only mode
Before, we were ignoring the "advance timelines" ticker, meaning we
would not initiate group commits. Normally, as part of group commit, we
call `advance_timelines`, which is what downgrades read holds.

Now, when in read-only mode, we instead call `advance_timelines`
directly, and make sure that it is aware of read-only mode.

Fixes MaterializeInc/database-issues#8940