storage: eagerly drop sink state in drop_sources_unvalidated
Same commit as the previous one, but for sinks.

Before, we were waiting on DroppedId messages for a couple of of things:
- cleaning up collection state in the controller
- updating mz_storage_shards (the global id -> shard id mapping)
- updating source/sink status to "dropped"
- removing statistics tracking

In some situations we don't get DroppedId messages, so we would not do
these updates/cleanups. The common of these situations are:
- dropping a replica right after dropping sources/sinks
- dropping a whole cluster

Now, we eagerly apply updates to our internal collections and also
eagerly clean up state.