compute/correction_v2: fix stage size tracking
This commit fixes a bug in size metrics tracking for the `CorrectionV2`
stage. In `consolidate_before`, we would call `stage.flush()` and if
this returned no updates we would assume that the stage had been empty
and immediately return without updating the metrics. However,
`stage.flush()` performs a `consolidate`, which can cancel out updates,
and we would miss recording these deletions in the metrics.

This fix chosen here is to perform the emptiness checking and early
return before calling `stage.flush()`. I find this more obvious than to
keeping the current flow and adding a second `update_metrics` to
`consolidate_before`, but both would work.