Force tokio thread names to fit Linux limits
At the moment, we cannot distinguish tokio thread names in Linux because
they're longer than the limit, cutting off their end. With this change,
we do the same we do in environmentd, which is to supply a custom thread
name function and construct the builder manually.

Updates `num_cpus` to the latest version.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>