compute: use epochs to distinguish client connections
Now that compute's internal command channel isn't recreated for each
client connection anymore, we need something else to distinguish client
connections and know when a reconciliation must be performed. To this
end, the `ClusterClient` is modified to pass through `CreateTimely`
commands to the workers. The compute worker is modified to monitor the
command stream for `CreateTimely` commands and initiate a reconciliation
when it observes one.

Additionally, the adapter thread that reads messages out of the
`ClusterClient`'s per-connection channels and pushes them into the
`Worker`'s persistent channel, and vice versa, is made aware about the
current connection's epoch to filter out responses not intended for that
connection.