compute: use epochs to distinguish client connections
Now that compute's internal command channel isn't recreated for each
client connection anymore, we need something else to distinguish client
connections and know when a reconciliation must be performed. To this
end, the channel adapter keeps an epoch that it increases on each client
reconnect and uses it to tag commands sent through the command channel.
The compute worker learns to monitor for epoch changes and initiate a
reconciliation when it observes one.

Additionally, the compute worker is modified to tag responses with the
current epoch and the channel adapter learns to filter out responses
with outdated epochs, which were intended for previous client
connections.