storage: when envd reconnects, report current status at latest timestamp
This makes it so that status updates from the latest deployment, for
example when coming out of read-only mode, sort last in
mz_source/sink_status_history, so get surfaced as the latest status in
mz_source/sink_statuses.

This fixes (or makes less likely) a bug where a failing/stalled
source/sink in the old deployment would crowd out the running status of
objects in the new deployment.

This fix work, but doesn't feel ideal. A problem is that the raw status
collections are append-only collections, so `envd` will simply append
status updates to the end. One could argue that the smarts should be in
`envd`/the controller. But ultimately, it would also have to append a
status update that "trumps" the earlier updates, so would do the same
thing that we're achieving with this change.

Fixes MaterializeInc/database-issues#8863