cluster: replace ClusterStartupEpoch with a nonce
The `ClusterStartupEpoch` is used in the Timely connection setup code to
ensure that inter-process connections are established correctly. That
code assumes that the epoch increases across controller reconnects, but
in a read-only environment, this isn't true.

This is an attempt at fixing the issue on the cluster side by using a
unique nonce per connection, instead of an increasing epoch. This makes
the Timely connection code lose the ability to determine which peer in
an incompatible connection attempt is executing an outdated iteration of
the cluster protocol and should shut down. The theory is that this is
fine if neither of the peers shuts down, because the outdated one will
eventually learn about the controller reconnect and re-instantiate the
Timely connection attempt with the correct nonce.