Logical dataflow backpressure
Implement a minimal approach to logical dataflow backpressure. We only
hold back capabilities of sources (not index imports), and only for data
that is not in advance of the upper at the time of installing the
dataflow, and release all capabilities once the input frontier advances
to the empty frontier. This is to avoid regressions where joins might
otherwise block compaction for longer than necessary.

The limitation might mean the feature is not effective, but it is the
only safe way to implement it as far as I can see.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>