cluster/communication: add turmoil test for create_sockets
This commit adds a chaos simulation for `create_sockets` using the
turmoil library. The test runs a number of processes and instructs
them to connect to each other, while simulating multiple controller
reconnects with increasing epochs.

Turmoil simulates the network, so it requires the use of its own
`TcpStream` types within the test. To enable that, support for the
turmoil TCP types is added to `mz-ore::netio`.

Run the test once with `cargo test -p mz-cluster`. Or use it for fuzzing
with `cargo test -p mz-cluster -- --ignored fuzz`.

Note that the test is currently failing unreliable. Dropping the
`create_sockets` future in a process can make other processes hang
attempting to connect to that process. It's unclear whether this is a
bug in turmoil or something we should expect given the absence of
network timeouts in `create_sockets`.