Configurable metrics interval for mz-metrics
Make the metrics collection interval configurable for the mz-metrics
crate. Previously, the interval was constant, which prevented changes at
runtime. However, we noticed that collecting lgalloc metrics can consume
significant CPU time, and we'd like to have the option to throttle, or
disable it.

The change adds a histogram for the metric update duration.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>