Update lgalloc to 0.5, expose new parameters
Updates the lgalloc dependency to 0.5, and exposes lgalloc's new parameters
as dyncfgs. Specifically, it allows us to set the local_buffer_bytes and
the file_growth_dampener. The first is the size of the per-thread per-size-
class buffer, the second is a dampener controlling how files per size class
grow as we need more capacity.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>