orchestrator: pass peer addresses to `args` closure
This commit re-introduces the `ServiceAssignments` type as the type
that's passed to the `args` closure provided in service configs. This
lets the `args` closure access not only the listen addresses of the
current process, but also the listen addresses of its peers, which will
be required for passing the Timely config on the process command line.