cluster*: pass Timely config through command line arguments
This config changes the cluster service creation code in the controller
to pass the Timely config into clusterd processes through command line
arguments.

The change is behind a feature flag: If
`enable_timely_init_at_process_startup` is not set (the current
default), the Timely config arguments are not passed.

Since the feature flag and some of the Timely config parameters are
provided as dyncfgs, this commit also gives the controller access to its
own `dyncfgs` value, like the compute/storage sub-controller already
have.