cluster: make create_sockets work without an external epoch
This commit adds a new version `create_sockets`, which doesn't require
an external epoch to be available. The new version is similar to the old
one in that it prevents processes from different generations from
connecting to each other. It achieves that using a pseudo epoch based on
the system time and a more sequential connection protocol that uses
process 0 as the generation leader.

The new version is added in a separate module, `communication_v2`, and
the old one is kept around. We want to keep it as a fallback for a
while, until the new version has proven itself in production.