cluster: make create_sockets work without an external epoch
This commit adapts `create_sockets` to work at cluster startup where
there is no external epoch available. The new version is similar to the
old one in that it prevents processes from different generations from
connecting to each other. It achieves that using a pseudo epoch based on
the system time and a more sequential connection protocol that uses
process 0 as the generation leader.