read max_credit_consumption_rate from a provided license key
if no license key is provided, default to a limit of 24 (cloud will use
a special license key which disables the check entirely)

if we have loaded a max_credit_consumption_rate, don't allow it to be
modified by anyone. this should ensure that we can continue managing
max_credit_consumption_rate as before in cloud (which will use a
specially issued license key which disables these checks) but anyone
using the emulator/self-hosted will either have a custom limit loaded
from their license key, or a default limit if they don't provide a
license key, neither of which can be overridden.