storage: add tests for storage controller metrics
This commit adds a cluster test for Prometheus controller metrics
exported by the storage controller, after the image of the existing
compute controller metrics test. It also fixes two dumb bugs in tracking
storage command history metrics.

Note that the final test for metrics cleanup fails currently, because we
fail to clean up non-primary tables and MVs correctly. This is fixed in
the next commit.