mv-sink: spread append work between all workers
This commit converts the MV sink's `append` operator from a
single-worker operator into a multi-worker operator, to share the work
of appending batches, as well as maintaining the output persist shard,
between all workers. The goal is to avoid skew in multi-process replicas
where one process has to perform an outsized amount of work compared to
the others.

The implementation extends the `BatchDescription` type by an
`append_worker` field that specifies the worker responsible for
appending the described batch and is chosen by the `mint` operator in
round-robin fashion. The `append_worker` index is then used in the
`append` operator to (a) route incoming batches to the responsible
workers and (b) ignore batch descriptions meant for other workers.

Note that we broadcast all batch descriptions to all `append` workers,
rather than routing them to the responsible worker only. We do this so
that workers can aggressively clean up their state when a batch
description is superseded by a new one (e.g. because a competing replica
managed to perform an append faster).