doc: add rbac note in dbt blue/green guide (#31842)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->

### Motivation
Adding a small note to mention the required permissions when using RBAC.