[persist] Flush inline parts to a single blob when possible (#31781)
The "tricky part" is that our append API accepts multiple batches, but
we need to generate a new part (or set of parts) that contain a subset
of the data from all of them, and has a lifecycle that many not match
any of them.

I considered moving the batch-merging to the caller when it is required,
but that requires some changes to the sink logic that increased the
blast radius of this change. (Probably still a good idea, but harder to
timebox.)

We instead deal with this by creating a new batch with a lifeycle bound
by the individual compare-and-append call, stuffing all our inline parts
in there, and flushing that batch instead.

### Motivation

Partially addresses:
https://github.com/MaterializeInc/incidents-and-escalations/issues/191

Discussed here:
https://materializeinc.slack.com/archives/C08ACQNGSQK/p1741192786064389?thread_ts=1741117192.727559&cid=C08ACQNGSQK

### Tips for reviewer

This ends up messier than I'd hoped, since we need to decode and
re-encode the data, and our current APIs for that have quite a bit of
migration cruft. I'm hoping to revisit the `FetchedPart` etc. hierarchy
once we've got structured data fully rolled out, which should make this
code a little less gnarly.

I've also added a dyncfg to derisk. (Default-on.)

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] If this PR includes major [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note),
I have pinged the relevant PM to schedule a changelog post.