doc: add note about Aurora PostgreSQL 16.1 logical replication bug (#31845)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->

### Motivation

Adding a small note about a known issue with Aurora PostgreSQL 16.1 that
affects logical replication:

>
https://repost.aws/questions/QU4RXUrLNQS_2oSwV34pmwww/error-could-not-map-filenumber-after-aurora-upgrade-to-16-1

I've wrapped this in a warning tag, but I'm open to suggestions if we
should use a different formatting approach or tag type for this
information.