storage: synthesize source-global 'paused' status updates for subsources
Before, we were only synthesizing this for the "main" ingestion, even
though we also get regular status updates for subsources (tables). Now
we synthesize a source-global status update for both, when all replicas
are dropped.

NOTE: This is in addition to the per-replica 'paused' updates that we're
also synthesizing when dropping a replica. We do this, so that we get a
source-global 'paused' status even when there are currently zero
replicas when creating a sources.